History task acomplished

+++ 23.1News/Models/Db/ExchangeRateEntity.cs
@@ -11,6 +11,6 @@ namespace _23._1News.Models.Db
         public DateTimeOffset? Timestamp { get; set; } = DateTimeOffset.Now;
         public ETag ETag { get; set; } = ETag.All;
         public string Currency { get; set; }
-        public decimal Rate { get; set; }
+        public double Rate { get; set; }
     }
 }
